+++ tests/test_utils.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import sys
 import warnings
 from pathlib import Path
@@ -9,7 +8,6 @@ import click
 import httpx
 import keyring.errors
 import pytest
-import shellingham
 from pydantic_core import Url
 
 import secrets_env.utils as t
@@ -164,45 +162,3 @@ class TestSetupCaptureWarnings:
             warnings.warn("test warning", UserWarning, stacklevel=1)
 
         assert "UserWarning: test warning" in caplog.text
-
-
-class TestDetectShell:
-    def test_success(self):
-        shell, path = t.detect_shell()
-        assert isinstance(shell, str)
-        assert isinstance(path, Path)
-
-    @pytest.mark.skipif(
-        os.getenv("CI") is None, reason="Expect GitHub Action environment"
-    )
-    def test_success__ci(self):
-        shell, path = t.detect_shell()
-        assert shell == "bash"
-        assert path == Path("/usr/bin/bash")
-
-    @pytest.mark.skipif(
-        os.getenv("SHELL") != "/bin/zsh", reason="This is a hard-coded test for zsh"
-    )
-    def test_success__local(self):
-        """Expect local shell is zsh"""
-        shell, path = t.detect_shell()
-        assert shell == "zsh"
-        assert path == Path("/bin/zsh")
-
-    def test_fallback(self, monkeypatch: pytest.MonkeyPatch):
-        monkeypatch.setattr(t, "_detect_shell_via_shellingham", lambda: None)
-        monkeypatch.setenv("SHELL", "/bin/sh")
-        shell, path = t.detect_shell()
-        assert shell == "sh"
-        assert path == Path("/bin/sh")
-
-    def test_shellingham_not_installed(self, monkeypatch: pytest.MonkeyPatch):
-        monkeypatch.setattr("builtins.__import__", Mock(side_effect=ImportError))
-        assert t._detect_shell_via_shellingham() is None
-
-    def test_shellingham_error(self, monkeypatch: pytest.MonkeyPatch):
-        monkeypatch.setattr(
-            "shellingham.detect_shell",
-            Mock(side_effect=shellingham.ShellDetectionFailure),
-        )
-        assert t._detect_shell_via_shellingham() is None
